refactor: TCA 아키텍처 적용 완료
- 즐겨찾기 구현
- 파일, 코드 컨벤션 정리
