+++ CleanArea/CleanArea/iOS/Detail/Components/DetailEtcView.swift
