+++ CleanArea/CleanArea/iOS/Detail/Components/DetailQualificationView.swift
