+++ CleanArea/CleanArea/iOS/Detail/Components/DetailWayView.swift
