+++ CleanArea/CleanArea/iOS/Enum/StartViewType.swift
@@ -5,8 +5,6 @@
 //  Created by 최동호 on 3/25/24.
 //
 
-import Foundation
-
 enum StartViewType {
     case startView
     case loading
