+++ CleanArea/CleanArea/iOS/Enum/TabType.swift
@@ -5,8 +5,6 @@
 //  Created by 최동호 on 12/11/23.
 //
 
-import Foundation
-
 enum TabType {
     case hot
     case like
