+++ CleanArea/CleanArea/iOS/Helpers/Modifiers/DetailModifer.swift
@@ -5,7 +5,6 @@
 //  Created by 최동호 on 12/7/23.
 //
 
-import Foundation
 import SwiftUI
 
 struct TagModifier: ViewModifier {
