+++ CleanArea/CleanArea/iOS/Main/View/ListView.swift
@@ -7,7 +7,6 @@
 
 import ComposableArchitecture
 
-import SwiftData
 import SwiftUI
 
 struct ListView: View {
