+++ CleanArea/CleanArea/iOS/Service/SwiftDataService.swift
@@ -14,29 +14,19 @@ import SwiftData
 struct SwiftDataService {
     @Query private var items: [YouthPolicy]
     
-    var fetch: () throws -> [YouthPolicy]
     var add: (YouthPolicy) throws -> Void
     var delete: (YouthPolicy) throws -> Void
+    var fetch: () throws -> [YouthPolicy]
     
     enum SwiftDataError: Error {
-        case fetchError
         case addError
         case deleteError
+        case fetchError
     }
 }
 
 extension SwiftDataService: DependencyKey {
     static let liveValue = Self(
-        fetch: {
-            do {
-                @Dependency(\.databaseService.context) var context
-                let itemContext = try context()
-                let descriptor = FetchDescriptor<YouthPolicy>()
-                return try itemContext.fetch(descriptor)
-            } catch {
-                throw SwiftDataError.fetchError
-            }
-        },
         add: { item in
             do {
                 @Dependency(\.databaseService.context) var context
@@ -53,6 +43,16 @@ extension SwiftDataService: DependencyKey {
                 itemContext.delete(item)
             } catch {
                 throw SwiftDataError.deleteError
+            }
+        },
+        fetch: {
+            do {
+                @Dependency(\.databaseService.context) var context
+                let itemContext = try context()
+                let descriptor = FetchDescriptor<YouthPolicy>()
+                return try itemContext.fetch(descriptor)
+            } catch {
+                throw SwiftDataError.fetchError
             }
         }
     )
